@@ -33,6 +33,8 @@
     "@material-ui/icons": "5.0.0-alpha.34",
     "@material-ui/lab": "5.0.0-alpha.34",
     "@material-ui/system": "5.0.0-alpha.34",
+    "@mui/icons-material": "^5.0.5",
+    "@mui/material": "^5.0.6",
     "@types/lodash": "^4.14.172",
     "@types/nprogress": "^0.2.0",
     "@types/react": "^17.0.19",
@@ -69,17 +71,22 @@
     "react-scripts": "^4.0.3",
     "react-syntax-highlighter": "^15.4.4",
     "stylis-plugin-rtl": "^2.1.0",
-    "typescript": "^4.2.4",
     "yup": "^0.32.9"
   },
   "devDependencies": {
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-typescript": "^12.3.1",
+    "@types/react-router": "^5.1.17",
+    "@typescript-eslint/eslint-plugin": "^5.2.0",
+    "eslint": "<8.0.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.0",
+    "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "prettier": "^2.3.2"
+    "husky": "^7.0.4",
+    "lint-staged": "^11.2.6",
+    "prettier": "^2.4.1",
+    "prettier-plugin-packagejson": "^2.2.13",
+    "prettier-plugin-sh": "^0.7.1",
+    "typescript": "^4.2.4"
   }
 }
