@@ -19,16 +19,14 @@ import {
 import MoreVertIcon from "@material-ui/icons/MoreVert";
 import { blue } from "@material-ui/core/colors";
 import { makeStyles } from "@material-ui/core/styles";
-import TimelineIcon from "@material-ui/icons/Timeline";
-
+import TableChartIcon from "@mui/icons-material/TableChart";
 import useSettings from "../../hooks/useSettings";
-import SpectrumPlot from "../../components/plots/SpectrumPlot";
-import mockSpectrumDataPayload from "../../mock/mockSpectrumData";
+import mockPhaseData from "../../mock/mockPhaseData";
+import SpectrogramTable from "src/components/plots/SpectrogramTable";
 
-const { REACT_APP_WS_USE_HOST, REACT_APP_WS_URL } = process.env;
-const hostURL = "ws://" + window.location.host + "/ws/consumer/spectrum";
-const spectrumAPI = REACT_APP_WS_USE_HOST ? hostURL : `${REACT_APP_WS_URL}/consumer/spectrum`;
-const ws = new WebSocket(spectrumAPI);
+const { REACT_APP_WS } = process.env;
+const phaseAPI = `${REACT_APP_WS}/consumer/phase`;
+const ws = new WebSocket(phaseAPI);
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -56,18 +54,66 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const Spectrum: FC = () => {
+const PhaseDisplay: FC = () => {
   const { settings } = useSettings();
   const classes = useStyles();
 
-  console.log("App: spectrumAPI: ", spectrumAPI);
+  console.log("PhaseDisplay: phaseAPI = ", phaseAPI);
 
-  const [data, setData] = useState([]); // mockSpectrumDataPayload.body
+  const [data, setData] = useState(null);
   const [socketStatus, setSocketStatus] = useState(Date().toLocaleString());
+
+  const spectrogramTable = new SpectrogramTable("phase-display-table");
+
+  //
+  // generate random data to test locally
+  //
+  /*
+  useEffect(() => {
+    let i = 1;
+    function myLoop() {
+      setTimeout(function () {
+        // randomly generated N = 100 length array 0 <= A[N] <= 39
+        const data1 = {
+          baseline: ["XX", "XY"],
+          polarisation: ["00", "01", "10"],
+          phase_values: [
+            [
+              Array.from({ length: 100 }, () => Math.floor(Math.random() * 360)),
+              Array.from({ length: 100 }, () => Math.floor(Math.random() * 270)),
+            ],
+            [
+              Array.from({ length: 100 }, () => Math.floor(Math.random() * 300)),
+              Array.from({ length: 100 }, () => Math.floor(Math.random() * 120)),
+            ],
+            [
+              Array.from({ length: 100 }, () => Math.floor(Math.random() * 180)),
+              Array.from({ length: 100 }, () => Math.floor(Math.random() * 90)),
+            ],
+          ],
+        };
+
+        setData(data1);
+        setSocketStatus(Date().toLocaleString());
+        console.log("data = ", data);
+
+        if (data1) spectrogramTable.draw(data1);
+
+        i++;
+        if (i < 10000) {
+          myLoop();
+        }
+      }, 1000);
+    }
+
+    myLoop(); //  start the loop
+  }, []);
+  */
 
   const onMessage = (event) => {
     const payload = JSON.parse(event.data);
-    console.log("App.js:onMessage: received payload = ", payload);
+    console.log("PhaseDisplay:onMessage: received event.data = ", event.data, typeof event.data);
+    console.log("PhaseDisplay:onMessage: received event.data = ", payload);
 
     if ("status" in payload) {
       console.log(payload.status);
@@ -77,27 +123,30 @@ const Spectrum: FC = () => {
     if ("body" in payload) {
       setData(payload.body);
       setSocketStatus(payload.timestamp);
+      spectrogramTable.draw(payload.body);
     }
   };
 
   useEffect(() => {
-    console.log("App: useEffect: 1");
+    console.log("PhaseDisplay: useEffect: 1");
+
     ws.onmessage = onMessage;
 
     return () => {
       // TODO
       // ws.close();
     };
-  });
+  }, []);
 
   useEffect(() => {
-    console.log("App: useEffect: 2");
+    console.log("PhaseDisplay: useEffect: 2");
+    // console.log("PhaseDisplay: data = ", JSON.stringify(data));
   }, [data, socketStatus]);
 
   return (
     <>
       <Helmet>
-        <title>Spectrum</title>
+        <title>Phase Display</title>
       </Helmet>
 
       <Box
@@ -119,15 +168,15 @@ const Spectrum: FC = () => {
                   }
                   avatar={
                     <Avatar className={classes.avatar}>
-                      <TimelineIcon />
+                      <TableChartIcon />
                     </Avatar>
                   }
-                  title="Spectrum Plot"
+                  title="Phase Display"
                   subheader={socketStatus}
                 />
 
                 <CardContent sx={{ pt: "8px" }}>
-                  <SpectrumPlot width={1000} height={500} data={data} />
+                  <div id="phase-display-table"></div>
                 </CardContent>
               </Card>
             </Grid>
@@ -138,4 +187,4 @@ const Spectrum: FC = () => {
   );
 };
 
-export default Spectrum;
+export default PhaseDisplay;
