@@ -1,17 +1,31 @@
-import logging
-
 import uvicorn
 from fastapi import FastAPI
 from starlette.middleware.gzip import GZipMiddleware
 from starlette.middleware.cors import CORSMiddleware
-
+import logging
+import logging.handlers
 import ska_ser_logging
 
 from server.core.settings import default_route_str
 from server.api import router as endpoint_router
 from server.core.config import PROJECT_NAME, BROKER_INSTANCE, LOGGING_LEVEL
 
-ska_ser_logging.configure_logging(LOGGING_LEVEL)
+ADDITIONAL_LOGGING_CONFIG = {
+    "handlers": {
+        "file": {
+            "()" : logging.handlers.RotatingFileHandler,
+            "formatter": "default",
+            "filename": "./output.log",
+            "maxBytes": 2048,
+            "backupCount": 2,
+        }
+    },
+    "root": {
+        "handlers": ["console", "file"],
+    }
+}
+
+ska_ser_logging.configure_logging(logging.INFO, overrides=ADDITIONAL_LOGGING_CONFIG)
 logger = logging.getLogger(__name__)
 
 app = FastAPI(title=PROJECT_NAME, version="2")
