@@ -107,41 +107,7 @@ const SettingsDrawer: FC = () => {
             ))}
           </TextField>
         </Box>
-        <Box
-          sx={{
-            mt: 2,
-            px: 1.5,
-          }}
-        >
-          <FormControlLabel
-            control={
-              <Switch
-                checked={values.direction === "rtl"}
-                color="primary"
-                edge="start"
-                name="direction"
-                onChange={(event): void =>
-                  handleChange(
-                    "direction",
-                    event.target.checked ? "rtl" : "ltr"
-                  )
-                }
-              />
-            }
-            label={
-              <div>
-                RTL
-                <Typography
-                  color="textSecondary"
-                  component="p"
-                  variant="caption"
-                >
-                  Change text direction
-                </Typography>
-              </div>
-            }
-          />
-        </Box>
+         
         <Box
           sx={{
             mt: 2,
