@@ -2,26 +2,23 @@ import * as d3 from "d3";
 import { useRef, useEffect } from "react";
 import PropTypes from "prop-types";
 
-const SpectrumPlot = (props) => {
+const Spectrum = (props) => {
   const ref = useRef();
 
   useEffect(() => {
-    d3.select(ref.current)
-      .attr("width", props.width)
-      .attr("height", props.height);
+    d3.select(ref.current).attr("width", props.width).attr("height", props.height);
     // .style("border", "1px solid black");
   }, []);
 
   function draw(width: number, height: number, data: any) {
-    console.log('SpectrumPlot:draw: data = ', data);
-    console.log('SpectrumPlot:draw: width, height =', width, height);
-    if (!data || !data.data || !width || !height) return;
+    console.log("SpectrumPlot:draw: data = ", data);
+    console.log("SpectrumPlot:draw: width, height =", width, height);
+    if (!data || !data.spectrum_values || !data.spectrum_values.length || !width || !height) return;
 
     // Clear
     d3.select(ref.current).select("svg").remove();
 
-    const { xMin, xMax, yMin, yMax, xLabel, yLabel } = data;
-    const values = data.data;
+    const { xMin, xMax, yMin, yMax, xLabel, yLabel, frequencies, rfis, flags, spectrum_values } = data;
 
     // set the dimensions and margins of the graph
     const margin = { top: 10, right: 10, bottom: 40, left: 50 };
@@ -39,10 +36,7 @@ const SpectrumPlot = (props) => {
 
     // Add X axis --> it is a date format
     const x = d3.scaleLinear().domain([xMin, xMax]).range([0, width]);
-    svg
-      .append("g")
-      .attr("transform", `translate(0,${height})`)
-      .call(d3.axisBottom(x));
+    svg.append("g").attr("transform", `translate(0,${height})`).call(d3.axisBottom(x));
 
     // Add Y axis
     const y = d3.scaleLinear().domain([yMin, yMax]).range([height, 0]);
@@ -51,7 +45,7 @@ const SpectrumPlot = (props) => {
     // Show confidence interval or std
     svg
       .append("path")
-      .datum(values)
+      .datum(spectrum_values)
       .attr("fill", "#1f77b4")
       .attr("stroke", "none")
       .attr("opacity", 0.3)
@@ -60,17 +54,17 @@ const SpectrumPlot = (props) => {
         d3
           .area()
           .curve(d3.curveMonotoneX)
-          .x((d) => x(d[0]))
-          .y0((d) => y(d[1] + d[2]))
-          .y1((d) => y(d[1] - d[3]))
+          .x((value, i) => x(frequencies[i]))
+          .y0((value) => y(value[0] + value[1]))
+          .y1((value) => y(value[0] - value[2])),
       );
 
     // Add the line
     svg
       .append("path")
-      .datum(values)
+      .datum(spectrum_values)
       .attr("fill", "none")
-      .attr("stroke", "#1f77b4")
+      .attr("stroke", "#3366CC")
       .attr("stroke-width", 1.5)
       .attr("opacity", 1)
       .attr(
@@ -78,9 +72,32 @@ const SpectrumPlot = (props) => {
         d3
           .line()
           .curve(d3.curveMonotoneX)
-          .x((d) => x(d[0]))
-          .y((d) => y(d[1]))
+          .x((value, i) => x(frequencies[i]))
+          .y((value) => y(value[0])),
       );
+
+    // rfi flags
+    if (rfis) {
+      svg
+        .selectAll(".point")
+        .data(spectrum_values)
+        .enter()
+        .append("path")
+        .attr("d", d3.symbol().type(d3.symbolSquare))
+        .style("fill", "#00CC66")
+        .style("opacity", (value, i) => flags[i] * 1)
+        .attr("transform", (value, i) => "translate(" + x(frequencies[i]) + "," + y(value[0]) + ")");
+
+      svg
+        .selectAll(".point")
+        .data(spectrum_values)
+        .enter()
+        .append("path")
+        .attr("d", d3.symbol().type(d3.symbolCircle))
+        .style("fill", "#CC3366")
+        .style("opacity", (value, i) => rfis[i] * 0.8)
+        .attr("transform", (value, i) => "translate(" + x(frequencies[i]) + "," + y(yMin) + ")");
+    }
 
     // Label for the x-axis
     svg
@@ -103,13 +120,7 @@ const SpectrumPlot = (props) => {
       .style("font-size", "15px")
       .text(yLabel);
 
-    svg
-      .exit()
-      .transition()
-      .duration(300)
-      .attr("y", height)
-      .attr("height", 0)
-      .remove();
+    svg.exit().transition().duration(300).attr("y", height).attr("height", 0).remove();
   }
 
   useEffect(() => {
@@ -124,10 +135,10 @@ const SpectrumPlot = (props) => {
   );
 };
 
-SpectrumPlot.propTypes = {
+Spectrum.propTypes = {
   width: PropTypes.number.isRequired,
   height: PropTypes.number.isRequired,
   data: PropTypes.any.isRequired,
 };
 
-export default SpectrumPlot;
+export default Spectrum;
