@@ -5,6 +5,12 @@ Functions to convert visibilities into QA plots.
 import numpy as np
 
 
+# translate numbers from a range to another range
+# by default, -180 - 180 -> 0 - 360
+def transfrom_hsl(x, min_a=-180, max_a=180, min_b=0, max_b=360):
+    return (x - min_a) * ((max_b - min_b) / (max_a - min_a)) + min_b
+
+
 def autospectrum_plot(data, antenna1, antenna2, frequency, chan_avg=5):
     """
     Generate plot of mean autospectrum in intensity over all antennas.
@@ -18,7 +24,8 @@ def autospectrum_plot(data, antenna1, antenna2, frequency, chan_avg=5):
     """
     # Extract autocorrelations and calculate their intensity
     auto = np.array(antenna1) == np.array(antenna2)
-    spectrum = 0.5 * (data[auto, :, 0] + data[auto, :, 3]).real
+    data_auto = data[auto, :, :]
+    spectrum = 0.5 * (data_auto[:, :, 0] + data_auto[:, :, 3]).real
 
     # Take mean over antennas
     spectrum = spectrum.mean(axis=0)
@@ -32,7 +39,7 @@ def autospectrum_plot(data, antenna1, antenna2, frequency, chan_avg=5):
     spec_std = tmp.std(axis=1)
 
     # Stack data into 2D array
-    data = np.stack((freq_avg, spec_avg, spec_std, spec_std)).T
+    data = np.stack((spec_avg, spec_std, spec_std)).T
 
     # Create body of plot
     body = {
@@ -43,7 +50,8 @@ def autospectrum_plot(data, antenna1, antenna2, frequency, chan_avg=5):
         "xMax": np.max(freq_avg),
         "yMin": np.min(spec_avg),
         "yMax": np.max(spec_avg),
-        "data": data.tolist(),
+        "frequencies": freq_avg.tolist(),
+        "spectrum_values": data.tolist(),
     }
 
     return body
@@ -73,13 +81,17 @@ def phase_plot(data, baseline, frequency, polarisation, chan_avg=50):
     # Get phase (in range -180 to 180)
     phase = np.angle(data_avg, deg=True)
 
+    ## convert phase (in range 0 to 360)
+    phase_hsl = transfrom_hsl(phase)
+
     # Create body of plot
     body = {
         "description": "Phase",
         "baseline": baseline,
         "frequency": freq_avg.tolist(),
         "polarisation": polarisation,
-        "phase": phase.tolist(),
+        ## (baseline, channel, polarisation) -> (baseline, polarisation, channel)
+        "phase_values": phase_hsl.transpose(0, 2, 1).tolist(),
     }
 
     return body
