@@ -1,25 +1,47 @@
 import asyncio
 import json
 import logging
-
 import uvicorn
 from aiokafka import AIOKafkaProducer
 from fastapi import FastAPI
-
 import ska_ser_logging
-
 from server.core.config import PROJECT_NAME, BROKER_INSTANCE, LOGGING_LEVEL
 from server.core.models.model import ProducerMessage, ProducerResponse
 
-ska_ser_logging.configure_logging(LOGGING_LEVEL)
+ADDITIONAL_LOGGING_CONFIG = {
+    "handlers": {
+        "file": {
+            "()" : logging.handlers.RotatingFileHandler,
+            "formatter": "default",
+            "filename": "./output.log",
+            "maxBytes": 2048,
+            "backupCount": 2,
+        }
+    },
+    "root": {
+        "handlers": ["console", "file"],
+    }
+}
+
+ska_ser_logging.configure_logging(logging.INFO, overrides=ADDITIONAL_LOGGING_CONFIG)
 logger = logging.getLogger(__name__)
 
 app = FastAPI(title=PROJECT_NAME)
 
 logger.info("PROJECT_NAME: %s; BROKER_INSTANCE: %s", PROJECT_NAME, BROKER_INSTANCE)
+
 loop = asyncio.get_event_loop()
+
+def serializer(value):
+    return json.dumps(value).encode("ascii")
+
 aioproducer = AIOKafkaProducer(
-    loop=loop, client_id=PROJECT_NAME, bootstrap_servers=BROKER_INSTANCE
+    loop=loop, 
+    client_id=PROJECT_NAME, 
+    bootstrap_servers=BROKER_INSTANCE, 
+    value_serializer=serializer,
+    compression_type="gzip",
+    max_request_size=15728640 # 15 MB
 )
 
 
@@ -48,10 +70,9 @@ async def broker_produce(msg: ProducerMessage):
     """
     payload = msg.dict()
     topic = payload.get("topic")
+    logger.info(f'broker_produce: payload = {payload}')
 
-    logger.info("broker_produce: payload = %s", payload)
-
-    await aioproducer.send(topic, json.dumps(payload).encode("ascii"))
+    await aioproducer.send(topic, payload)
 
     response = ProducerResponse(topic=topic)
     logger.info("broker_produce: response = %s", response)
